day 5
yeah... I tried to parse the initial state and that was a bad idea. But I managed to do it in the end

Go. I'm getting used to the syntax and I managed to gather some speed but

Apartenly Go cannot have a simple Min/Max function for ints. Either you do casts and your code is ugly or you have to implement your own Min/Max. WTF!!! I really disagree with the statement in Go documentation in which they say that method overloading is not that needed.

Doing AoC in go is like going back to Data Structures 101 and have to implement Stack, Heaps, etc.. Seriously? Of course, it took me ages to figure out the Stack code

Also error messages must be in lowercase. WTF

On the plus side, I'm loving the multiple returns. I wish I could have it on Java or php