day 7
today things were a ittle confused because on go you need to be aware of
the difference between a pointer and an struct.
Blessed Smalltalk (and Java who copied it) for making everything a pointer
and solving this mess in a way that you can save your brain power for other things.

Once I realize where to put the pointers things were easy. solution is not as clean as
I would like to, because I discovered the lambdas (not sure on the go naming for that)
too late to refactor