day9
No much to complain today.
The gomum library could had been more useful as I had to do manually half of the self growing matrix
An of course, why pointers? It's never clear when my changes modify the object or no. Going back to pointers
is making me appreciate Java a lot more.

The time wasted dealing with pointers (and it's errors) is far greater than the time spent on stupid Java boilerplate