day 15 (part 1 only)
painful day.
Approach 1: matrix and drawing points => out of memory creating the matrix
Approach 2: just keep the lines and the occupied points => the distance between the sensor and the beacon is far to much to check every point
Approach 3: calculate the range each sensor intercepts with the wanted line and reduce the ranges that overlap. Worked

Both approach 1 and 2 were working before I discovered they did not work on the input, so too much time wasted. But on the bright side approach one drawed a beatiful step by step (too much time wasted drawing it) that was useful to debug the other 2 approaches.

Part 2 is almost there. It seems that calculating 4 000 000 ranges is OK (a couple of minutes) so it only needs to find the empty point in the ranges and discard those points occupied by sensors or beacons

I had to refactor quite a lot of stuff from previous days, moving the Point and Lines from day 14 to be reused and added an extra string array parameter to the solver interface so I can add custom overrides for tests (eg the line for the smaller example)

go rant of the day:
I want the collections package. I wasted too much time removing elements from an slice (but on retrospective I should have used a map) and it's nice to have sets, I know, they're just maps on the inside but the code
it's clearer