day 19
I barely made it. Slow as hell and  for some reasons it fails with the sample input for part 2 but it worked on AoC
I'll debug it at some point in the future

At the beginning on they day I though immutables will make my life easy with the recursion. And I could not figure out how to
discard branches. I looked for some hints on reddit but even with that it took a couple of hours. I lost the output and I
rewrote everything not cloning objects and using good old school value parameters and now it completes is a reasonable time
(10 minutes or so)

I tried memoizing but it did not make much of a difference (to be fair, i was before I discover that cloning was the main problem)

I probably could have try to take advantage of go threading but I managed without it in the end

I could use the lessons from today to attempt day 16

Go rant...
The usual, constructors, verbose map creation and not having inheritance