day 21
an easy day. Using pointers and a tree made the recursion really easy. I could not reuse the model for part 2 but
I had a lot of fun implementing the equation solving algorithm that I remember from school.

go rant...
Inheritance would had made today really easy. But this half-baked interfaces in go are almost useless and painful to use.
You don't need to declare interfaces. That looks good on the paper, but it makes almost imposible to track the usages
and then it's difficult for your IDE to help you implementing the interface, but in got you have to copy/paste or type
because your cannot read your mind.

And for the n-th time. Inheritance is good. Today I had to add methods to the interface that I could have inherited and then implement
them on every subclass
Also I had to add methods to the interace that made no sense because you cannot cast back to the actual class
It was so so painful today. And all the madness with the interace pointer receiver's
This is bad. Allow me to cast freely like C or give me inheritance and runtime matching. But without any of thos is frustrating and
error-prone

day 17 (part 2 not working but some progress)

There is a pattern and it can be found with state+pos+shape but converting it to an actual solution doesn't seem to work. Also for some reason
even part1 does not seem to pass the tests anymore (but it still gives me the right solution for the input)
I know I should add the lines up to the pattern start but even with that the number is still too low.
I've attempted 1556521739125 and 1556521739122. Both too low