day 22 (part 1 only)
arg...another painfull days after spending hours to discover that the only problem was that the last movemnet have no rotation.
I went for a linked list approach for the edge transitions that once you solve the how to connect edges problem then the movemnet
it's piece of cake.
I think it was quite intentional that the last movement for the example was irrelevant.

part 2 (on it's way)
after getting my cardboard companion cube to calculate how to connect edges by naming them a have a matrix with the connections
(one matrix for the 2d another one for 3d) I discovered that the way the faces are folded in the input it's quite different
than for the example.

So I decided to define 2 maps with face coordinates and the face object will have the connection. I managed to get
part 1 working with that approach and I'm in the middle of connecting the 3d edges. I think I have direction rotation
on edge transition sorted and the only problmem it's that when connecting edge1(row) = edge2(row)
or edge1(column) = edge2(column) it's not true anymore, only for no rotation

go rant...
I want to have struct attributes or variables named map or type. Seriosly type is such a common attribute