day 17 (part2)
after many reviews of the vertex finally I managed to have the right map. Probably I could have done everything thinking
on face rotations... but it's done.
The map is more complex as it needs to track not only the face it's connected but the vertex. I introduced a new struct (connection)
and rewrote the connect methods to do the transformations required (depending on the destination vertex)
Also I needed to remove the bidirectional closing, but going through all the faces and borders eventually closes everything