day 25
piece of cake (compared with the last week)

go rant

quite disappointed with go. I cannot see any reasonable use case for it. It's just a better C, but with all the C annoyances
this half-baked OO is clearly insuficient to take advantage of OO

It doesn't have enought library power to compete with PHP and I rather use python half-baked OO. It's not good either but
it's much better than this

It's fast and light weight and the syntax it's OK but without OO or free cast it's really difficult to model

Generics seems a nice addition but they're quite recent to the language

tooling wise, the debugger is quite limited (I'm so used to be able to do proper evaulations) and you cannot have a quick varilable to
check a return function value. Because the compiler won't alow you to run with the unused varilable (to be fair probably that's something
that can be changed on the settings)
Maybe GoLand it's good, but VS Code go it's limited.

I don't think I will use it for anything, but I will definitelly repeat the experience of doing next year AoC in a language I want to
learn. It's definitelly how to learn a programming language in 25 days.

Next year it will be either Scala or Python (I need to learn python properly)