day 16
An absolute nightmare...

After rewriting it like day24 (priority queue + cache) I managed to get it working for part1. Initially was quite slow but I managed to
prune efficiently and got part 1 on a nice speed (because I know I need it for part 2)

It took me ages, partially because of some issues with pointers and go walking the maps randomly doesn't make debugging easy.

For part2 first I needed to rewrite everything just to have the open nodes in the state and not the path. After a lot of refactoring I managed to
have it working with just the open nodes

Then I struggled on how to model both paths, in the end I managed to have both possitions and both countdowns in the state and selecting 2
contributing valves each time

That worked for the small example but I couple of bugs prevent the real input to work, including one major bug in which I discarded the result
before checking if the output was bigger than the max it the state could not open more valves. Of course the solution was amongs those.

It runs too slow and I think it's adding to much options and repeating them, but at some point I decided to try the current max in AoC and that worked.
Not proud but i'm too tired of this problem.

Go rant...

so if you use & operator on the varilable created on an iterator, well you'll get a pointer to the same address. WTF

eg
key,v := range myMap {
  if v.output > max {
    max = valve.output
    maxValve = &v
  }

And you think that maxValve will hold a pointer to the max... No, it holds a pointer to the last used. Yep, that's fun an unpredictable

You need to do

  key,v := range myMap {
  if v.output > max {
    max = valve.output
    maxValve = &myMap[key]
  }

also on the subject of maps... what I mentioned before. When you iterates a map the order is random. Normally in other languages the order
is not guaranteed but if you put the same keys they will iterate in the same order. So when you debug and run, things are consistent. But here
every time you debug the same case to try to debug the issue mentioned above you have different executions every single time which delays the
debugging a lot

And on the top of thaat, I can still be ranting about the lack of ternaries or the ugly map notation...